# Use an explicit buffer?