There's a perf penalty for the simplest case, and benefits for the most complex ones. The code size doesn't move. The resulting string is also flat vs probably a tree-like rope for what the current version produces. The latter could be slower to parse by the DOM api (that overhead is not currently benchmarked).

Before:

```
Classcat – Strings × 6,655,176 ops/sec
classNames – Strings × 2,773,836 ops/sec
Fastest is Classcat – Strings

Classcat – Objects × 4,805,586 ops/sec
classNames – Objects × 2,317,781 ops/sec
Fastest is Classcat – Objects

Classcat – Strings & Objects × 4,259,464 ops/sec
classNames – Strings & Objects × 2,129,470 ops/sec
Fastest is Classcat – Strings & Objects

Classcat – Mixed × 2,571,844 ops/sec
classNames – Mixed × 1,564,550 ops/sec
Fastest is Classcat – Mixed

Classcat – Arrays × 2,082,569 ops/sec
classNames – Arrays × 669,301 ops/sec
Fastest is Classcat – Arrays
```

After:

```
Classcat – Strings × 5,564,369 ops/sec
classNames – Strings × 2,688,253 ops/sec
Fastest is Classcat – Strings

Classcat – Objects × 5,670,332 ops/sec
classNames – Objects × 2,343,242 ops/sec
Fastest is Classcat – Objects

Classcat – Strings & Objects × 5,375,772 ops/sec
classNames – Strings & Objects × 2,150,893 ops/sec
Fastest is Classcat – Strings & Objects

Classcat – Mixed × 5,521,760 ops/sec
classNames – Mixed × 1,553,639 ops/sec
Fastest is Classcat – Mixed

Classcat – Arrays × 5,573,590 ops/sec
classNames – Arrays × 585,414 ops/sec
Fastest is Classcat – Arrays
```