@JorgeBucaran 2018-02-12
@pygy This is pretty cool. Do you have any idea or could you explain why this change has such a perf boost?

Also, 

> The resulting string is also flat vs probably a tree-like rope for what the current version produces. 

What would be an example of a string that is a "tree-like rope"?