@pygy 2018-02-12
You can shave a few bytes by iterating backwards over arrays: `for (var i = classes.length; i--;) {` . No perf impact in node, and the class order in `class="..."` attributes doesn't have any impact on styles.