@JorgeBucaran 2018-02-12
@pygy I am only asking why using an array has better perf than concatenating as I am currently doing?