@pygy 2018-02-12
I hadn't seen your comment yet :-)

Re. perf boost: fewer branches and temp allocations.

For ropes, JS engine have many optimizations for strings, among which, when you concatenate two of them, rather than copying the bytes in a third one, it can create a [rope object](https://en.wikipedia.org/wiki/Rope_(data_structure)) (that appears as a string to JS). The C++ DOM code, when it extacts the classes, has to either flatten the rope before parsing, or walk that tree (not sure what it does).