@JorgeBucaran 2018-02-12
@pygy Thanks. Is there any way to avoid having two functions?