@pygy 2018-02-12
At the expense of extra branches... Let me try...