@pygy 2018-02-12
There, somehow, it is even faster :-)

Before:

```
Classcat – Strings × 6,694,269 ops/sec
classNames – Strings × 2,751,419 ops/sec
Fastest is Classcat – Strings

Classcat – Objects × 4,820,041 ops/sec
classNames – Objects × 2,345,441 ops/sec
Fastest is Classcat – Objects

Classcat – Strings & Objects × 4,183,793 ops/sec
classNames – Strings & Objects × 2,178,062 ops/sec
Fastest is Classcat – Strings & Objects

Classcat – Mixed × 2,553,804 ops/sec
classNames – Mixed × 1,580,940 ops/sec
Fastest is Classcat – Mixed

Classcat – Arrays × 2,064,071 ops/sec
classNames – Arrays × 652,851 ops/sec
Fastest is Classcat – Arrays
```

After:

```
Classcat – Strings × 6,281,381 ops/sec
classNames – Strings × 2,712,796 ops/sec
Fastest is Classcat – Strings

Classcat – Objects × 6,482,448 ops/sec
classNames – Objects × 2,299,174 ops/sec
Fastest is Classcat – Objects

Classcat – Strings & Objects × 6,431,835 ops/sec
classNames – Strings & Objects × 2,110,564 ops/sec
Fastest is Classcat – Strings & Objects

Classcat – Mixed × 6,327,718 ops/sec
classNames – Mixed × 1,485,107 ops/sec
Fastest is Classcat – Mixed

Classcat – Arrays × 6,509,578 ops/sec
classNames – Arrays × 675,852 ops/sec
Fastest is Classcat – Arrays
```