@pygy 2018-02-12
The latest push is purely cosmetic (variable naming: `out` => `buffer`).