@jorgebucaran 2018-08-18
Hi @pygy! I was going over this issue (again) and thought I'd include your changes in the next release. It really took me a while to come around this, so apologies for not getting back to your earlier. 

Turns out I never did run the benchmarks on your branch. But then I just _did_, and there must be a regression of some kind because the following numbers don't agree with what you wrote before. In fact, classcat came out behind classNames in some instances. 😮 

```console
# Strings
classnames × 4,183,036 ops/sec
classcat × 3,005,044 ops/sec

# Objects
classnames × 4,590,641 ops/sec
classcat × 3,389,816 ops/sec

# Strings and Objects
classnames × 3,895,419 ops/sec
classcat × 2,902,810 ops/sec

# Mixed
classnames × 2,929,269 ops/sec
classcat × 2,271,244 ops/sec

# Arrays
classnames × 1,021,356 ops/sec
classcat × 1,782,158 ops/sec
```