@pygy 2018-08-19
Hi Jorge, this is what I have today, using Node@10.1.0 and classnames@2.2.5:

```
Classcat – Strings × 7,862,926 ops/sec
classNames – Strings × 2,820,883 ops/sec
Fastest is Classcat – Strings

Classcat – Objects × 7,834,685 ops/sec
classNames – Objects × 3,097,185 ops/sec
Fastest is Classcat – Objects

Classcat – Strings & Objects × 7,649,889 ops/sec
classNames – Strings & Objects × 2,527,958 ops/sec
Fastest is Classcat – Strings & Objects

Classcat – Mixed × 7,731,479 ops/sec
classNames – Mixed × 2,014,684 ops/sec
Fastest is Classcat – Mixed

Classcat – Arrays × 7,932,852 ops/sec
classNames – Arrays × 703,937 ops/sec
Fastest is Classcat – Arrays
```

I swapped the lib order in the benchmark, to do it like you (I've had surprises in the past where changing the order would yield different results du to JIT feebly heuristics), same result:

```
classNames – Strings × 2,755,935 ops/sec
Classcat – Strings × 7,601,639 ops/sec
Fastest is Classcat – Strings

classNames – Objects × 2,992,971 ops/sec
Classcat – Objects × 7,508,957 ops/sec
Fastest is Classcat – Objects

classNames – Strings & Objects × 2,567,248 ops/sec
Classcat – Strings & Objects × 7,429,014 ops/sec
Fastest is Classcat – Strings & Objects

classNames – Mixed × 1,991,416 ops/sec
Classcat – Mixed × 7,468,549 ops/sec
Fastest is Classcat – Mixed

classNames – Arrays × 690,681 ops/sec
Classcat – Arrays × 7,575,605 ops/sec
Fastest is Classcat – Arrays
```